price list generation in Queue

+++ app/Providers/AppServiceProvider.php
@@ -8,6 +8,7 @@ use Illuminate\Queue\Events\JobProcessed;
 use Illuminate\Queue\Events\JobProcessing;
 use Illuminate\Queue\Events\JobFailed;
 use App\JobStatus;
+use App\Jobs\ParsePrice;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -29,36 +30,50 @@ class AppServiceProvider extends ServiceProvider
     public function boot()
     {
         Queue::before(function (JobProcessing $event) {
-            $payload = $event->job->payload();
-            $job = unserialize($payload['data']['command']);
 
-            JobStatus::updateOrCreate(
-                ['contractor_id' => $job->getContractorId()],
-                [
-                    'status_id' => 2,
-                    'message' => 'Прайс в процессе обработки',
-                ]
-            );
+            $class = $event->job->resolveName();
+
+            if (is_a($class, ParsePrice::class, true)) {
+                $payload = $event->job->payload();
+                $job = unserialize($payload['data']['command']);
+
+                JobStatus::updateOrCreate(
+                    ['contractor_id' => $job->getContractorId()],
+                    [
+                        'status_id' => 2,
+                        'message' => 'Прайс в процессе обработки',
+                    ]
+                );
+            }
+
         });
 
         Queue::after(function (JobProcessed $event) {
-            $payload = $event->job->payload();
-            $job = unserialize($payload['data']['command']);
+            $class = $event->job->resolveName();
 
-            JobStatus::where(['contractor_id' => $job->getContractorId()])->delete();
+            if (is_a($class, ParsePrice::class, true)) {
+                $payload = $event->job->payload();
+                $job = unserialize($payload['data']['command']);
+
+                JobStatus::where(['contractor_id' => $job->getContractorId()])->delete();
+            }
         });
 
         Queue::failing(function (JobFailed $event) {
-            $payload = $event->job->payload();
-            $job = unserialize($payload['data']['command']);
+            $class = $event->job->resolveName();
 
-            JobStatus::updateOrCreate(
-                ['contractor_id' => null],
-                [
-                    'status_id' => 3,
-                    'message' => 'При обработке прайса произошла ошибка!',
-                ]
-            );
+            if (is_a($class, ParsePrice::class, true)) {
+                $payload = $event->job->payload();
+                $job = unserialize($payload['data']['command']);
+
+                JobStatus::updateOrCreate(
+                    ['contractor_id' => $job->getContractorId()],
+                    [
+                        'status_id' => 3,
+                        'message' => 'При обработке прайса произошла ошибка!',
+                    ]
+                );
+            }
         });
 
         Queue::looping(function () {
